[mapleall]Internal commit msg:     [cgemit] emit visibility for every external variable     [ICSL] Clean Code 13     do not duplicate debug scope in cfgo

+++ src/mapleall/maple_be/src/cg/reg_alloc_lsra.cpp
@@ -1949,7 +1949,7 @@ void LSRALinearScanRegAllocator::LiveIntervalAnalysis() {
           }
           lowestLi->SetStackSlot(kSpilled);
           active.erase(itFinded);
-          interNum = active.size();
+          interNum = static_cast<uint32>(active.size());
         } else {
           break;
         }
